+++ app/room/joinPrivateRoom.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { FormEvent, useState } from "react";
+import { FormEvent } from "react";
 
 interface JoinPrivateRoomProps {
   privateRoomIds: String[];
